fix(frontend): wrap auth pages with suspense and adjust signup formatting
